Added initial support for generalized arrays and domains.  Arrays and domains are supported currently up to rank three but can easily be extended to any rank with few changes---additional access macros in the array macros runtime header and additional "this" functions in the internal prelude.
Added a pass to InsertAnonymousTypes.  This handles array and domain
types put on variables.  Given
  var D : domain(1) = 1..n;
  var A : [D] integer;
It puts the following in commonModule for use by any modules.
  type _domain_1d = domain(1);
  type _array_integer_1d = [D] integer;
It does this when the array's type is a primitive.  If the array type
is defined by the programmer, the array type declaration follows that
declaration.

Note that the type of D is part of the type of _array_integer_1d even
though D is not.

Removed functionality of 2D arrays of floats and 1D arrays of integers
and 1D and 2D domains from the runtime files.  These are now generated
in the compiler.

Added a test case for 3D array of integers.  Currently, arrays do not
seem to work if the elements are assigned values other than zero.

The flowanalysis tests changed because of the change to the internal
prelude and because array and domain types are now in the compiler.
They have associated type symbols with real names.

There is a fair amount of cleanup and generalization work left to do
but this checkin constitutes a significant portion of array/domain
support.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1476 3a8e244f-b0f2-452b-bcba-4c88e055c3ca