Fixed a bug in arrays of higher dimensions having trouble storing their values that I introduced in the last checkin.
Added a test case that stores values other than zero in an array.
Unfortunately this test case fails in analysis (future for John)
because the index variables used in the forall expression are added
and assigned to the array of integers.  Can the index type be treated
symmetrically to integers?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1480 3a8e244f-b0f2-452b-bcba-4c88e055c3ca