Added a field parentType to the Type class.  This field will store a subtype relationship for a Type; The Type is a subtype of its parentType.  This field is currently the nilType in all cases.
Added two tuple future tests.  Failing in analysis and beyond
analysis.

Improved parser support for tuples.  They can now be used on the left
hand side of assignment statements.  This resulted in the following
two changes to the Spec.

1. The statement in a function definition must be one of the following:
  A compound statement ({ ... })
  An empty statement (;)
  A conditional statement (if ...)
  A for, forall, or while loop (for ..., forall ..., while ..., do ...)
  A call statement (call ...)
  A return statement (return ...)

Rationale. I originally made it just a block statement to get around
the parser ambiguities.  David suggested it could also be a return
statement.  I added the rest (at least in the implementation for now).

The ambiguity stems from starting assignment statements with tuples
clashing with the return type (domains and indexes) of the function:
  function foo() : domain(expr);
  function foo() : domain(i,j) = (k,l);
Types that can be qualified by expressions clash with the tuples
starting the next statement in one-lookahead parsers.

2. The while-, for-, forall-, and if-statements now require a "do" or
"then" to avoid using curly brackets around the bodies of the loops or
the conditional "then" clause.

The ambiguity can be seen in the example
  while foo(a,b) = (j,k);
Is foo a function call with actuals a and b or a boolean with a tuple
assignment as the body.  Of course, again, we know the answer, but
we're using a one-lookahead parser.

Changed test cases to reflect language changes.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1490 3a8e244f-b0f2-452b-bcba-4c88e055c3ca