Made ForallExpr store indices as Symbol* rather than Stmt* of VarDefStmts.  This gets rid of the Stmts in Exprs.  Made defPoint in Symbol point to BaseAST rather than Stmt since it may point to either an Expr or a Stmt now.
Made Fixup-verify once again check to make sure that all defPoints are
correct.

In coping with this change, modified analysis to see whether the
defPoint is an Expr or a Stmt before grabbing the ainfo.  This is a
tad verbose but seems necessary.  The other option is to put ainfo and
asymbol in BaseAST and have them only active for Expr/Stmt and
Symbol/Type respectively.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1512 3a8e244f-b0f2-452b-bcba-4c88e055c3ca