Added wrappers to the memory functions malloc, calloc, realloc, and free.  At present the only benefit is error checking with hopefully descriptive error messages, but we have great plans for these functions in the future.
The arguments to _chpl_malloc, _chpl_calloc, and _chpl_realloc
differ from the originals.  Instead of being passed the size
of the entire chunk of memory being requested they take the
number of elements and then the size of the element.
_chpl_malloc and _chpl_calloc also take a string which
is a description of the type of memory requested.

Example:
--------
  arr.base = (elemtype*)_chpl_malloc(arr.size, sizeof(elemtype), "array")

Corresponding error message:
----------------------------
  ***Error:  Out of memory allocating array***


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1514 3a8e244f-b0f2-452b-bcba-4c88e055c3ca