Improved copy resolution.  (2/16 TO DO #1) This checkin changes the way in which copied symbols are re-resolved after copy is called.
At the end of copy, when called on an Expr or Stmt, a new traversal
called UpdateSymbols runs through the new Expr or Stmt, replacing
Variable references to old symbols with variable references to the new
symbols (if the symbol was copied).

In support of this, copying of Variable expressions no longer make its
symbol unresolved.  They keep the symbol the same.

This new way of handling copying things like block statements (things
that need their scopes copied) is much more sane.

Removed functions call_cleanup and call_cleanup_ls which used to be
called by copy.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1520 3a8e244f-b0f2-452b-bcba-4c88e055c3ca