Fixed a couple issues with class cloning and resolve symbols. test/types/type_variables/deitz/test_clone_class1 is now working correctly. Updated .future files for test/types/type_variables/deitz/test_clone_class2(a) since I fixed a couple issues.  Here is the new .future:
"The problem here is that classes have already been cloned, but
the original class (from which the clones were derived) still
exists in the AST via the TypeDefStmt.  Like unused functions
there will be unused types which should not be resolved.

If you need an interface to query which types are actually
used in the program, ask me."


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1537 3a8e244f-b0f2-452b-bcba-4c88e055c3ca