Replaced the Cleanup traversal with the following traversals:
  ExpandClassWiths
  InsertThisParameters
  ScopeResolveSymbols
  RenameOverloadedFunctions
  SpecializeParenOpExprs
  ApplyThisParameters

These traversals used to be embedded in Cleanup, but now show up in
the Passlist.  This makes the design slightly more modular and
uniform.

Removed four lines that fix the back pointer after Cleanup was called
on copied expressions and statements.  Since Cleanup is no longer
called there, these are no longer necessary.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1543 3a8e244f-b0f2-452b-bcba-4c88e055c3ca