Moved TransformLetExprs after analysis.
This breaks test_let[1-7].  I added futures for each of these tests
for John.

Note to whom-will-be-concerned: I'm not letting them regress (adding
futures instead), because this will require new functionality.  It was
premature to remove the futures before.  Therefore, these are not
regressions, but rather not yet fully implemented.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1551 3a8e244f-b0f2-452b-bcba-4c88e055c3ca