Changed the contents of TypeDefStmt from a Type to a TypeSymbol.  This brings it more inline with FnDefStmt, VarDefStmt, and ModuleDefStmt. All of these contain the symbols they define.
Reworked TypeSymbol::clone so that it creates a new name for the type
symbol.  Previously this was hacky and done in the copy mechanism.

There is a deep difference between TypeSymbol::clone used to clone
classes and FnSymbol::clone used to clone functions.  When a class is
cloned, the name is changed.  When a function is cloned, the name is
not changed, the clone is treated like any other overloaded function,
and the cname is changed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1597 3a8e244f-b0f2-452b-bcba-4c88e055c3ca