Fixed last night's valgrind regression, which related to complex literals that were expressed using integer sub-values.  The problem was related to an idiom in my coding that I started out using, but have since shied away from (though haven't completely eradicated yet).  In particular, I used to do:
	if (typeid(foo) == typeid(DesiredType)) {
	  DesiredType foo2 = (DesiredType)(foo);
	  ...
	}

using a static rather than a dynamic cast, having proved it was
safe to do so.  I started to worry that this was a fragile
construct since someone might modify the conditional without
modifying the code.  It turns out that in this case, I did just
that.  So now I use dynamic casts inside of these conditionals
as an extra safety check (and performance be damned! :)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1611 3a8e244f-b0f2-452b-bcba-4c88e055c3ca