Added a new pass InsertFunctionTemps that inserts temporaries for cases where multiple functions are called from the same statement. This makes sure we use the right order of evaluation and makes it easier to manipulate function calls.  This maybe should be changed eventually to deal only with side-affecting functions.
Added scope functions findEnclosingSymContext,
findEnclosingStmtContext, and findEnclosingExprContext to search up
from a scope looking for contexts.  Cleaned up code that did the same
thing by hand.

Updated typeName array so that print_view works again.  I should add
something to make sure this stays up-to-date when new AST classes are
added or removed.

Fixed, with a stopgap, the check in fixup for the parentSymbol
pointer.  This is now lenient allowing either the module or initFn to
be the parent for top-level code.  The __init function should be
inserted and then fixed up like other things.

Added two new tests to look at deeply nesting function calls as in

  foo(foo(foo(foo(foo(2)))))
  foo(foo(foo(foo(foo(2) + foo(2)) + foo(2)) + foo(2)) + foo(2)) + foo(2)

Updated test_var_decl_init1 to obey new parsing rules for
if-statements.  Namely, if-statements must use a 'then' keyword or use
curly brackets.

Removed futures for trivial/deitz/test_var_decl_init1 and
trivial/deitz/test_let8.  Note that the first test only works because
of the fntemps pass breaking up two function calls in a statement.
They are not side-affecting so don't have to be broken up.  The
ProcessParameters pass fails in that case though.  This is still a
problem.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1651 3a8e244f-b0f2-452b-bcba-4c88e055c3ca