Trying to ease back into implementation, I worked on getting parameter intents working correctly for class arguments.  This was mostly fairly straightforward -- I ammended the logic that determines how formal arguments should be codegened and how blank parameters are interpreted to handle the class cases correctly (in particular, blank intents imply pass-by-reference for classes).
This fix also revealed some subtle problems in call_fixup
within stmt.cpp which Steve helped me correct.

I'm checked in four tests: 2 work correctly; 1 fails
in analysis because it uses equality (==) on class
instances, which seems to trip analysis up.  This is
a future I've assigned to John;  a second test fails
because it tries to refer to "nil" which we haven't
implemented yet.  I've put that one in my name.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1654 3a8e244f-b0f2-452b-bcba-4c88e055c3ca