A few changes to make the compiler work on the Sun boxes again.  Added an explicit cast to files.cpp, made some of the commands executed by the test scripts more portable.
The one change that might affect users is that the
test environment variable CHPL_VALGRIND_OPTIONS was
shortened to CHPL_VALGRIND_OPTS because the former name
was too long for the Sun boxes.  John, I think you might
have used this at one point when we had a version skew?
I'm not sure whether you still do or not.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1703 3a8e244f-b0f2-452b-bcba-4c88e055c3ca