Enabled BuildClassConstructorsEtc which currently just builds an == function for records.
Made it so that RemoveDeadSymbols now removes dead functions as well
as dead types right after analysis.

Added a guard against codegening an empty list of statements for a
module.  Added a guard against codegening constructors if there are
none.  (For records, the constructor might be dead and thus removed.)

Made a couple of superficial changes to BuildClassConstructorsEtc.

Added a new test test_record_equal1 to test if we can use == on records.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1737 3a8e244f-b0f2-452b-bcba-4c88e055c3ca