Changed to:
deitz

record of record is not working

The problem here is that buildClassConstructorsEtc is
not calling the constructor for the embedded record.
Analysis is using the constructor call to setup the
type information, so even if we don't want to make
a full "__new" call we could make an "__init" call
with the embedded type to tell analysis what it is.
Basically I would like a hook so that I can initialize
the type of "coordinates".

I am passing this back to Steve to make the call
to the sub-constructor.  If we want to handle this
some other way, let's talk about it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1757 3a8e244f-b0f2-452b-bcba-4c88e055c3ca