Added codegenPrototype routines for arrays, classes, records, and unions.  This allows these to be declared after they are used.  This currently shows up because of the way arrays are created but could show up if the user declares them after they are used.
Improved support for arrays of records/classes/unions.  The test for
array of records test is currently failing in analysis, but the test
of an array of classes passes.

Fixed a bug in the _INIT_ARRAY macro which would fail if the domain or
array were named "d" because it used a local variable named d as well.
This local variable is now called "__dim".


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1798 3a8e244f-b0f2-452b-bcba-4c88e055c3ca