Added some initial support for I/O on classes (and types in general) via bound functions.  This isn't at all ready for prime-time yet, but I got enough done today (and it's stable enough) that I want to check in.  The changes were:
1) Added a new SymtabTraversal insertWriteFns that
traverses the Symboltable looking for classes that
don't have write functions declared.  When it
finds one, it creates a write function for it
(albeit one that does nothing useful other than
prove to me that it's been created -- since these
are created after analysis, they are considered
dead by default for now and don't get codegenerated).

2) Refactored the I/O code so that it's more type-
oriented -- rather than handling I/O for all types
in IOExpr, pushed this code into the Type hierarchy
for simplicity.  Eventually, it'll probably be
removed from here as well, but for the time being,
this made it easier to deal with.

3) Added a boolean to the SymtabTraversal pass
that allows you to skip over "internal" scopes,
which include the intrinsic scope, the prelude
scopes, and the common scope.  This is similar
to the controls we already have in the normal
Traversals to avoid certain parts of the program.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1855 3a8e244f-b0f2-452b-bcba-4c88e055c3ca