Added initial support for sequences (literal sequence expressions and cruftily printing sequences).  Added new Expr subclass SeqExpr and new Type subclass SeqType which are sequence literals and sequence types respectively.
SeqType is a subclass of ClassType.  This allows us to implement it as
a class.  The strategy for sequences is to create it as a class and to
create a node class.  This is implemented using linked lists at the
AST level.  Note that once class cloning and class type variables are
working we may be able to leverage that for our sequence
implementation too.

Added a new traversal expandSeqExprAssignments that expands
assignments of literal sequences into multiple append lines.  Append
is currently a runtime function but will eventually be a method in the
AST of SeqType.

Initialization of SeqTypes works the same as any ClassType by using
the constructor.  The defaultVal for SeqTypes is a call to the
constructor not nil.  However, the defaultVal for the node's class is
still nil.

Added new future to test parsing 'seq(integer)' in addition to 'seq of
integer'.

Added new sub-directory to my seq tests called noanalysis which uses
the --noanalysis flag.  These tests now work without analysis so I'm
putting this there so I can maintain them until they work with
analysis.

Plugged CommonModule with a stopgap measure to not print out static in
front of its functions during codegen because these can be used in
other modules.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1867 3a8e244f-b0f2-452b-bcba-4c88e055c3ca