Implemented default I/O for classes, records, and unions. These currently print out as follows:
record myrecord {
  var c: integer;     =>     (c=1, d=1.2)
  var d: float;
}

union myunion {
  var e: integer;     =>     (uninitialized)
  var f: float;       OR:    (e=1)
}                     OR:    (f=1.2)

class myclass {
  var a: integer;     =>     nil
  var b: float;       OR:    {a=1, b=1.2}
}

Feedback on these print formats would be appreciated; they
can easily be changed.  The record format was chosen because
it seemed it was so close to the record literal format; since
unions are also value types but can only have one active field
at a time, a similar format was used there, with the somewhat
ugly "(uninitialized)" standing in for a union that has never
been assigned.  The class format was meant to be distinguishable
from the record format but similar.  Another approach would be
to generate them as myclass(a=1, b=1.2) so it looked more like
a constructor...

The intention is to support the user's ability to over-ride
these defaults by defining their own bound write() function,
but this doesn't work yet.  Once it does, the default format
matters somewhat less as long as its clear and succinct enough
that people might use it when they're being lazy.

These functions are currently added after analysis has run
so that we can be sure about the types of everything.  Eventually,
we will probably want to move this before analysis and convert
IOCalls into dispatches to write() methods in order to have
analysis include the write() methods as part of its dead
code analysis.

Added a bunch of tests, many of which are futures, but I'll
mail people about those separately.

This check-in includes a few incidental changes:

- added a new runtime union macro -- _UNION_CHECK_QUIET()
  that performs the union safe read check without generating
  an error if it's mismatched.  This can be used to create
  if-then-else's that check for all the union's cases
  without generating errors.  Also rewrote _UNION_CHECK()
  to use _UNION_CHECK_QUIET() in its definition and to
  write its error to stderr rather than stdout.

- added a new state to unions which indicates that they
  are uninitialized so that we can distinguish this case
  from a union whose memory just happens to be initialized
  to zero.  Interestingly, I was guessing that this would
  result in a valgrind error, as I don't think we currently
  set a union's enum to 0, yet it didn't...  ?

- refactored Steve's codegenIOCall() for sequences to be a
  method of sequence rather than a dynamic cast within
  ClassType

- added a method that supports the ability to tack on an
  "else" clause to a CondStmt after it has been constructed
  in order to build cascading if-then-else's easily.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1921 3a8e244f-b0f2-452b-bcba-4c88e055c3ca