Reassigned and annotated .future files:
OK, I took a look at the results with ANALYSIS_MATCH turned on, and
they look correct to me.  We get one "warning" at line 78, but that
is for "f" being set to a cloned type (which it should be), and
we get one warning at line 112 which is the return type of the cloned
construct_foo function which is now returning the new cloned type
(which it should be).

*** note: the problem is that the AST's types are not specific, sure
    the variable is a "foo", but it is really a clone of "foo".
    Perhaps ANALYSIS_MATCH needs to be enhanced.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1948 3a8e244f-b0f2-452b-bcba-4c88e055c3ca