Fixed a bug in Expr's call_fixup used when expressions are replaced.
The defaultVal of record types and union types is now a call to the
default constructor.  This causes the constructor to be called on all
records as they are created.  However, analysis doesn't see this and
marks the constructor as dead.  This has resulted in a few tests that
were seemingly working now not working.  Since they were never
working, I refuturized them.  John, let me know if you have any
questions.  Basically, the constructor call is in the defaultVal
expression of a type which is used by default to initialize variables.
It is therefore not dead.

Arrays now insert default and user initializations when they are
created.  So arrays no longer have to be initialized if the default
initialization will do.

I special-cased a test for inserting the default initialization
statement for classes within their constructor.  Since we declare a
"this" that will be return in the constructor, it is important not to
initialize it.  Surprisingly, doing so crashed my system multiple
times.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1959 3a8e244f-b0f2-452b-bcba-4c88e055c3ca